+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Features
 
+* add `CreatePayment` `GetPayment` `ListPayments`
+
 <a name-"v0.7.1"></a>
 
 ## [v0.7.1](https://github.com/linode/linodego/compare/v0.7.0..v0.7.1) (2018-02-05)